requirements.txt
openai
python-dotenv
requests
python-telegram-bot